back_end/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
numpy<2.0.0
fastapi
uvicorn[standard]
python-multipart
openai>=1.0.0
pandas
python-dotenv
gtts
requests
openai-whisper
ffmpeg-python
openpyxl
librosa
soundfile
pydub 